requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
environs==9.5.0
markdown==3.3.6
psycopg2-binary==2.9.3
redis==4.3.1
requests==2.27.1
